refactor: trace response in IT
