+++ mesh-integration-tests/src/test/java/com/netcracker/it/common/HttpClient.java
@@ -11,9 +11,10 @@ import java.util.concurrent.TimeUnit;
 
 @Slf4j
 public class HttpClient {
+
     public static OkHttpClient okHttpClient = new OkHttpClient.Builder()
             .addInterceptor(withRetryOnServiceUnavailableOrTimeout())
-            .readTimeout(130, TimeUnit.SECONDS)
+            .readTimeout(60, TimeUnit.SECONDS)
             .connectTimeout(60, TimeUnit.SECONDS)
             .sslSocketFactory(TlsUtils.getSslContext().getSocketFactory(), TlsUtils.getTrustManager())
             .build();
