--- mesh-integration-tests/src/test/java/com/netcracker/it/go/model/TraceResponse.java
@@ -1,18 +1,0 @@
-package com.netcracker.it.go.model;
-
-import lombok.Data;
-
-@Data
-public class TraceResponse {
-    String serviceName;
-    String familyName;
-    String version;
-    String podId;
-
-    String requestHost;
-    String serverHost;
-    String remoteAddr;
-    String path;
-    String method;
-    String xversion;
-}
