+++ mesh-integration-tests/src/test/java/com/netcracker/it/quarkus/HttpIT.java
@@ -4,7 +4,7 @@ import com.google.gson.Gson;
 import com.netcracker.cloud.junit.cloudcore.extension.annotations.*;
 //import com.netcracker.cloud.junit.cloudcore.extension.client.KubernetesClientFactory;
 //import com.netcracker.cloud.junit.cloudcore.service.ITHelper;
-import com.netcracker.it.quarkus.model.TraceResponse;
+import com.netcracker.it.common.model.TraceResponse;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.junit.jupiter.api.BeforeAll;
