+++ mesh-integration-tests/src/test/java/com/netcracker/it/spring/CompositeIT.java
@@ -5,8 +5,8 @@ import com.netcracker.cloud.junit.cloudcore.extension.annotations.Cloud;
 import com.netcracker.cloud.junit.cloudcore.extension.annotations.EnableExtension;
 import com.netcracker.cloud.junit.cloudcore.extension.annotations.PortForward;
 import com.netcracker.cloud.junit.cloudcore.extension.annotations.Value;
+import com.netcracker.it.common.model.TraceResponse;
 import com.netcracker.it.spring.model.CompositeStructure;
-import com.netcracker.it.spring.model.TraceResponse;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import okhttp3.HttpUrl;
 import okhttp3.Request;
